requirements.txt
requests~=2.27.1
environs~=9.5.0
python-telegram-bot==13.11
